Fix a concurrency issue with atomics

+++ core/connection.go
@@ -110,7 +110,7 @@ type Connection struct {
 	msgMapByPath map[string]map[uint16]api.Message // map of messages indexed by message ID which are indexed by path
 
 	channelsLock  sync.RWMutex        // lock for the channels map and the channel ID
-	nextChannelID uint16              // next potential channel ID (the real limit is 2^15)
+	nextChannelID uint32              // next potential channel ID (the real limit is 2^15)
 	channels      map[uint16]*Channel // map of all API channels indexed by the channel ID
 
 	subscriptionsLock sync.RWMutex                  // lock for the subscriptions map
